Move api key handling to own module
