+++ server/src/api-keys/api-keys.controller.spec.ts
@@ -1,0 +1,20 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { ApiKeysController } from './api-keys.controller';
+import { ApiKeysService } from './api-keys.service';
+
+describe('ApiKeysController', () => {
+  let controller: ApiKeysController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [ApiKeysController],
+      providers: [ApiKeysService],
+    }).compile();
+
+    controller = module.get<ApiKeysController>(ApiKeysController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
