+++ server/src/auth/auth.controller.ts
@@ -6,17 +6,8 @@ import {
   Post,
   UseGuards,
   Request,
-  Param,
-  Delete,
 } from '@nestjs/common';
 import { ApiBearerAuth } from '@nestjs/swagger';
-import { ApiKeysService } from './api-keys.service';
-import {
-  CreateApiKeyRequestDto,
-  CreateApiKeyResponseDto,
-  GetApiKeyResponseDto,
-  GetApiKeysResponseDto,
-} from './api-keys.dto';
 import {
   LoginRequestDto,
   LoginResponseDto,
@@ -29,10 +20,7 @@ import { LocalAuthGuard } from './local-auth.guard';
 
 @Controller('auth')
 export class AuthController {
-  constructor(
-    private authService: AuthService,
-    private apiKeysService: ApiKeysService,
-  ) {}
+  constructor(private authService: AuthService) {}
 
   @UseGuards(LocalAuthGuard)
   @Post('login')
@@ -56,42 +44,5 @@ export class AuthController {
   @Get('me')
   getMyProfile(@Request() req: any) {
     return req.user;
-  }
-
-  @UseGuards(JwtAccessAuthGuard)
-  @ApiBearerAuth()
-  @Get('api-keys')
-  @HttpCode(200)
-  getAllApiKeys(): GetApiKeysResponseDto {
-    return {
-      apiKeys: this.apiKeysService
-        .getAllKeyNames()
-        .map((keyName): GetApiKeyResponseDto => {
-          return {
-            keyName: keyName,
-          };
-        }),
-    };
-  }
-
-  @UseGuards(JwtAccessAuthGuard)
-  @ApiBearerAuth()
-  @Post('api-keys')
-  @HttpCode(200)
-  createApiKey(
-    @Body() createApiKeyDto: CreateApiKeyRequestDto,
-  ): CreateApiKeyResponseDto {
-    return {
-      keyName: createApiKeyDto.keyName,
-      apiKey: this.apiKeysService.create(createApiKeyDto.keyName),
-    };
-  }
-
-  @UseGuards(JwtAccessAuthGuard)
-  @ApiBearerAuth()
-  @Delete('api-keys/:keyName')
-  @HttpCode(200)
-  deleteApiKey(@Param() params: { keyName: string }): void {
-    return this.apiKeysService.delete(params.keyName);
   }
 }
