+++ server/src/auth/auth.module.ts
@@ -3,7 +3,6 @@ import { ConfigModule } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { UsersModule } from '../users/users.module';
-import { ApiKeysService } from './api-keys.service';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { JwtAccessStrategy } from './jwt-access.strategy';
@@ -17,7 +16,6 @@ import { LocalStrategy } from './local.strategy';
     LocalStrategy,
     JwtAccessStrategy,
     JwtRefreshStrategy,
-    ApiKeysService,
   ],
   controllers: [AuthController],
 })
