+++ server/test/auth.e2e-spec.ts
@@ -105,64 +105,4 @@ describe('AuthController (e2e)', () => {
         .expect(401);
     });
   });
-
-  describe('/auth/api-keys (POST)', () => {
-    it('should be protected', async () => {
-      await request(app.getHttpServer())
-        .post('/auth/api-keys')
-        .send({
-          keyName: 'testkey',
-        })
-        .expect(401);
-    });
-
-    it('should create api keys', async () => {
-      const accessToken = await getAccessToken();
-      const apiKeyResponse = await request(app.getHttpServer())
-        .post('/auth/api-keys')
-        .auth(accessToken, { type: 'bearer' })
-        .send({
-          keyName: 'testkey',
-        })
-        .expect(200);
-      expect(apiKeyResponse.body.apiKey).toBeDefined();
-    });
-  });
-
-  describe('/auth/api-keys (GET)', () => {
-    it('should be protected', async () => {
-      await request(app.getHttpServer()).get('/auth/api-keys').expect(401);
-    });
-
-    it('should list api keys', async () => {
-      const accessToken = await getAccessToken();
-      await request(app.getHttpServer())
-        .post('/auth/api-keys')
-        .auth(accessToken, { type: 'bearer' })
-        .send({
-          keyName: 'testkey',
-        })
-        .expect(200);
-
-      const apiKeyResponse = await request(app.getHttpServer())
-        .get('/auth/api-keys')
-        .auth(accessToken, { type: 'bearer' })
-        .expect(200);
-      expect(apiKeyResponse.body.apiKeys).toEqual([
-        {
-          keyName: 'testkey',
-        },
-      ]);
-    });
-  });
-
-  async function getAccessToken(): Promise<string> {
-    const loginResponse = await request(app.getHttpServer())
-      .post('/auth/login')
-      .send({
-        username: 'myuser',
-        password: '12345',
-      });
-    return loginResponse.body.access_token;
-  }
 });
